+++ src/Veneer/Binding.php
@@ -125,7 +125,7 @@ class Binding
                     if($needsSlingshot) {
                         if(!class_exists(Slingshot::class)) {
                             throw Exceptional::ComponentUnavailable(
-                                'Cannot resolve constructor dependencies without Slingshot'
+                                message: 'Cannot resolve constructor dependencies without Slingshot'
                             );
                         }
 
@@ -366,7 +366,9 @@ class Binding
         ContainerProvider $containerProvider
     ): void {
         if ($this->proxy === null) {
-            throw Exceptional::Setup('Target binding has not been created');
+            throw Exceptional::Setup(
+                message: 'Target binding has not been created'
+            );
         }
 
         $instRef = new ReflectionClass($instance);